Added manage account panel. Updated onboarding prefab
