+++ Scuti/Resources/Materials/RoundedBanner.mat
@@ -82,4 +82,4 @@ Material:
     m_Colors:
     - _Color: {r: 1, g: 1, b: 1, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
-    - _WidthHeightRadius: {r: 0, g: 84, b: 50, a: 0}
+    - _WidthHeightRadius: {r: 672, g: 84, b: 50, a: 0}
