+++ Scuti/Scripts/UI/Account/ManagerAccount.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 927004ed17a5f334db6013608729fc70
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
